requirements.txt
absl-py==0.8.0
aiofiles==0.4.0
aiohttp==3.6.0
APScheduler==3.6.1
asn1crypto==0.24.0
astor==0.8.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.7
blis==0.2.4
boto3==1.9.228
botocore==1.12.228
bz2file==0.98
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.14.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
dill==0.3.0
docopt==0.6.2
docutils==0.15.2
dopamine-rl==2.0.5
en-core-web-sm==2.1.0
fbmessenger==6.0.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
gevent==1.4.0
gin-config==0.2.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-pasta==0.1.7
googleapis-common-protos==1.6.0
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.9.0
gym==0.14.0
h5py==2.10.0
httplib2==0.13.1
httptools==0.0.13
humanfriendly==4.18
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.2
jsonschema==2.6.0
kafka-python==1.4.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mattermostwrapper==2.1
mesh-tensorflow==0.0.5
mpmath==1.1.0
mtranslate==1.6
multidict==4.5.2
murmurhash==1.0.2
networkx==2.3
nltk==3.4.5
numpy==1.17.2
oauth2client==4.1.3
opencv-python==4.1.1.26
packaging==19.1
pandas==0.25.1
paramiko==2.6.0
pika==1.0.1
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pydot==1.4.1
pyglet==1.3.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pypng==0.0.20
PySocks==1.7.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-engineio==3.9.3
python-socketio==4.3.1
python-telegram-bot==11.1.0
pytz==2019.2
PyYAML==5.1.2
questionary==1.3.0
rasa==1.3.3
rasa-sdk==1.3.2
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.34
rsa==4.0
ruamel.yaml==0.15.100
s3transfer==0.2.1
sanic==19.3.1
Sanic-Cors==0.9.9.post1
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.8.2
scikit-learn==0.20.4
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
spacy==2.1.8
SQLAlchemy==1.3.8
srsly==0.1.0
sshtunnel==0.1.5
sympy==1.4
tabulate==0.8.3
tensor2tensor==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-datasets==1.2.0
tensorflow-estimator==1.14.0
tensorflow-gan==1.0.0.dev0
tensorflow-metadata==0.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.0.8
tqdm==4.35.0
twilio==6.30.0
typing-extensions==3.7.4
tzlocal==2.0.0
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.3
uvloop==0.13.0
wasabi==0.2.2
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.15.6
word2number==1.1
wrapt==1.11.2
xlrd==1.2.0
yarl==1.3.0

